Minor bug fixes

+++ src/olaf_max_filter_lemire.c
@@ -74,7 +74,7 @@ void olaf_lemire_max_filter(float* array, size_t array_size , size_t filter_widt
 }
 
 void olaf_max_filter_naive(float* array, size_t array_size , size_t filter_width , float* maxvalues){
-    size_t half_filter_width = filter_width / 2;
+    int half_filter_width = filter_width / 2;
     for(size_t i = 0 ; i < array_size; i++){
         size_t startIndex = i - half_filter_width > 0 ? i - half_filter_width : 0;
         size_t stopIndex = i + half_filter_width < array_size ? i + half_filter_width + 1: array_size;
